+++ packages/cloud_firestore/test/cloud_firestore_test.dart
@@ -96,6 +96,7 @@ void main() {
               'path': 'foo',
               'parameters': <String, dynamic>{
                 'where': <List<dynamic>>[],
+                'orderBy': <List<dynamic>>[],
               }
             },
           ),
@@ -124,6 +125,7 @@ void main() {
                   'where': <List<dynamic>>[
                     <dynamic>['createdAt', '<', 100],
                   ],
+                  'orderBy': <List<dynamic>>[],
                 }
               },
             ),
@@ -153,6 +155,7 @@ void main() {
                   'where': <List<dynamic>>[
                     <dynamic>['profile', '==', null],
                   ],
+                  'orderBy': <List<dynamic>>[],
                 }
               },
             ),
@@ -180,7 +183,9 @@ void main() {
                 'path': 'foo',
                 'parameters': <String, dynamic>{
                   'where': <List<dynamic>>[],
-                  'orderBy': <dynamic>['createdAt', false],
+                  'orderBy': <List<dynamic>>[
+                    <dynamic>['createdAt', false]
+                  ],
                 }
               },
             ),